requirements.txt
aioredis==1.3.1
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.0
constantly==15.1.0
cryptography==41.0.3
daphne==3.0.2
Django==4.0.5
django-cors-headers==4.2.0
django-filter==23.2
djangochannelsrestframework==1.0.0
djangorestframework==3.13.1
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
mysql-connector-python==8.1.0
mysqlclient==2.2.0
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
pyOpenSSL==23.2.0
pytz==2023.3
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
zope.interface==6.0
